fix options
The options passed to xattr are internally converted from XATTR_XATTR_*
to XATTR_* with the exception of NOFOLLOW which is in most cases handled
explicitly in the code.

In order to make this work properly, however, the correct constants need
to actually be used in python - that means the XATTR_XATTR_* ones must
be there rather than the OS's XATTR_* ones.