Make tests independent of user's locale
As os.fsencode() is not present in Python 2.7, a custom wrapper is
used which calls sys.getfilesystemencoding() to detect filesystem
encoding. This detection is based on user's locale settings, so
in different environments it could be either Unicode or non-Unicode value,
and the tests will behave differently in this case.

To make tests independent of that, explicitly set 'compat.fs_encoding'
variable and run tests for both Unicode and ASCII.