Note about how to create custom xattr
I'm not sure if this is quite the whole truth of it, but a note like this would've helped me avoid some confusion.

On Ubuntu 20.1 with EXT4, this differs from the `attr` command, which automatically prefixes the key with `user.`.  So, if I add an xattr with the linux command `-s "testkey" "value" /path/to/file', I have to read from xattr with the key `user.testkey`.